# apidoc/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CatCafeApi",
    "description": "猫カフェのポータルサイトと管理画面で使用するAPIを定義",
    "version": "0.1.0"
  },
...
  "servers": [
    {
      "url": "http://{username}:{port}/{basePath}",
      "description": "ローカル環境",
      "variables": {
        "username": {
          "default": "localhost"
        },
        "port": {
          "enum": [
            "8080"
          ],
          "default": "8080"
        },
        "basePath": {
          "default": ""
        }
      }
    }
  ],
  "paths": {
...
    "/{cats}/contact": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "post": {
        "tags": [
          "contact"
        ],
        "summary": "お問い合わせ登録API",
        "description": "お問い合わせを1件登録する",
        "operationId": "contactCreate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/contact/request.json#/definitions/contact_create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/postResponse.json#/definitions/post_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/contacts": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "get": {
        "tags": [
          "contact"
        ],
        "summary": "お問い合わせ一覧取得API",
        "description": "お問い合わせを取得する",
        "operationId": "contactSearch",
        "parameters": [
          {
            "name": "store_ids",
            "in": "query",
            "description": "店舗ID",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/contact/response.json#/definitions/contact_search_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/contacts/{contact_id}": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "contact_id",
          "in": "path",
          "description": "お問い合わせID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "get": {
        "tags": [
          "contact"
        ],
        "summary": "お問い合わせ詳細取得API",
        "description": "お問い合わせ詳細を1件取得する",
        "operationId": "contactFind",
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/contact/response.json#/definitions/contact_find_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          }
        }
      }
    },
...
    "/{cats}/casts": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "get": {
        "tags": [
          "cast"
        ],
        "summary": "キャスト一覧取得API",
        "description": "キャストを取得する",
        "operationId": "castSearch",
        "parameters": [
          {
            "name": "store_ids",
            "in": "query",
            "description": "店舗ID",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "取得サイズ",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/cast/response.json#/definitions/cast_search_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/stores/{store_id}/cast": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "store_id",
          "in": "path",
          "description": "店舗ID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "post": {
        "tags": [
          "cast"
        ],
        "summary": "キャスト登録API",
        "description": "キャストを1件登録する",
        "operationId": "castCreate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/cast/request.json#/definitions/cast_create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/postResponse.json#/definitions/post_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/stores/{store_id}/casts/{cast_id}": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "store_id",
          "in": "path",
          "description": "店舗ID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        },
        {
          "name": "cast_id",
          "in": "path",
          "description": "キャストID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "get": {
        "tags": [
          "cast"
        ],
        "summary": "キャスト詳細取得API",
        "description": "キャスト詳細を1件取得する",
        "operationId": "castFind",
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/cast/response.json#/definitions/cast_find_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          }
        }
      },
      "patch": {
        "tags": [
          "cast"
        ],
        "summary": "キャスト更新API",
        "description": "キャストを1件更新する",
        "operationId": "castUpdate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/cast/request.json#/definitions/cast_update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/patchResponse.json#/definitions/patch_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      },
      "delete": {
        "tags": [
          "cast"
        ],
        "summary": "キャスト削除API",
        "description": "キャストを1件論理削除する",
        "operationId": "castDelete",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "description": "バージョンフィールド",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "正常系"
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      }
    }
  }
}